style: page 내부 component 폴더의 파일들을 다 폴더로 만드로 index로 접근하게 수정
