+++ src/pages/AddPage/components/MemoTextAreaWithAIButton/index.tsx
@@ -1,0 +1,1 @@
+export { MemoTextAreaWithAIButton } from './MemoTextAreaWithAIButton';
