+++ src/pages/AddPage/components/ResultMemoList/index.tsx
@@ -1,0 +1,1 @@
+export { ResultMemoList } from './ResultMemoList';
