+++ src/pages/DashboardPage/components/SelectedTagMemosList/index.tsx
@@ -1,0 +1,1 @@
+export { SelectedTagMemosList } from './SelectedTagMemosList';
