+++ src/pages/SearchPage/components/SearchInput/index.tsx
@@ -1,0 +1,1 @@
+export { SearchInput } from './SearchInput';
