+++ src/pages/SearchPage/components/SearchScrollView/AIMessage.tsx
